Switch KisPainter to use floating point opacity instead of 0...255 integer range
Globally, we store opacity as a floating point value. It was quite weird that we
converted the value to quint8, when passing to KisPainter, then back to float when
passing to KoCompositeOp and then, finally, to channels_type.

Now KisPainter operates with floating-point opacity value. It removes all these
extra conversions and errors when rounding.

BUG:356462

+++ libs/image/kis_indirect_painting_support.h
@@ -42,7 +42,7 @@ public:
     virtual void setCurrentColor(const KoColor &color);
     void setTemporaryTarget(KisPaintDeviceSP t);
     void setTemporaryCompositeOp(const QString &id);
-    void setTemporaryOpacity(quint8 o);
+    void setTemporaryOpacity(qreal o);
     void setTemporaryChannelFlags(const QBitArray& channelFlags);
     void setTemporarySelection(KisSelectionSP selection);
