+++ libs/ui/processing/KisEncloseAndFillProcessingVisitor.h
@@ -42,7 +42,7 @@ public:
         bool unmerged,
         bool useBgColor,
         bool useCustomBlendingOptions,
-        int customOpacity,
+        qreal customOpacity,
         const QString &customCompositeOp,
         QSharedPointer<QRect> outDirtyRect
     );
@@ -76,7 +76,7 @@ private:
     bool m_unmerged {false};
     bool m_useBgColor {false};
     bool m_useCustomBlendingOptions;
-    int m_customOpacity;
+    qreal m_customOpacity;
     QString m_customCompositeOp;
     QSharedPointer<QRect> m_outDirtyRect;
 };
